predictor/requirements.txt
numpy
pandas
xgboost
google-cloud-aiplatform